feat(lib/optee): introduce API get_optee_header_ep
get_optee_header_ep() returns the OP-TEE pager entrypoint from
a loaded op-tee image if found. This information helps platform to
locate OP-TEE load area when platform support paged and unpaged OP-TEE
configuration in which cases OP-TEE is loaded in a specific memory
as an internal memory when pager is enabled or the external memory
when pager is disabled.

Change-Id: Idc7dde091f2ada8898f40d02e68c3834ca39d8e8
Signed-off-by: Etienne Carriere <etienne.carriere@st.com>
Signed-off-by: Yann Gautier <yann.gautier@foss.st.com>